44 super_admin and admin access for school, staff

+++ private/controllers/Schools.php
@@ -13,10 +13,15 @@ class Schools extends Controller
         $crumbs[] = ['Schools', 'schools'];
 
         $data = $school->findAll();
-        $this->view('schools', [
-            'rows' => $data,
-            'crumbs' => $crumbs
-        ]);
+
+        if(Auth::access('super_admin')) {
+            $this->view('schools', [
+               'rows' => $data,
+               'crumbs' => $crumbs
+            ]);
+        }else{
+            $this->view('access-denied');
+        }
     }
 
     public function add() {
@@ -26,7 +31,7 @@ class Schools extends Controller
 
         $errors = array();
 
-        if(count($_POST) > 0) {
+        if(count($_POST) > 0 && Auth::access('super_admin')) {
             $school = new School();
 
             if($school->validate($_POST)) {
@@ -43,10 +48,14 @@ class Schools extends Controller
         $crumbs[] = ['Schools', 'schools'];
         $crumbs[] = ['Add', 'schools/add'];
 
-        $this->view('schools.add', [
-            'errors' => $errors,
-            'crumbs' => $crumbs
-        ]);
+        if(Auth::access('super_admin')) {
+            $this->view('schools.add', [
+                'errors' => $errors,
+                'crumbs' => $crumbs
+            ]);
+        }else{
+            $this->view('access-denied');
+        }
     }
 
     public function edit($id = null) {
@@ -58,7 +67,7 @@ class Schools extends Controller
 
         $school = new School();
 
-        if(count($_POST) > 0) {
+        if(count($_POST) > 0 && Auth::access('super_admin')) {
             if($school->validate($_POST)) {
                 $school->update($id, $_POST);
                 $this->redirect('schools');
@@ -73,11 +82,15 @@ class Schools extends Controller
         $crumbs[] = ['Schools', 'schools'];
         $crumbs[] = ['Edit', 'schools/edit'];
 
-        $this->view('schools.edit', [
-            'row' => $row,
-            'errors' => $errors,
-            'crumbs' => $crumbs
-        ]);
+        if(Auth::access('super_admin')) {
+                $this->view('schools.edit', [
+                'row' => $row,
+                'errors' => $errors,
+                'crumbs' => $crumbs
+            ]);
+        }else{
+            $this->view('access-denied');
+        }
     }
 
     public function delete($id = null) {
@@ -89,7 +102,7 @@ class Schools extends Controller
 
         $school = new School();
 
-        if(count($_POST) > 0) {
+        if(count($_POST) > 0 && Auth::access('super_admin')) {
             $school->delete($id);
             $this->redirect('schools');
         }
@@ -100,10 +113,13 @@ class Schools extends Controller
         $crumbs[] = ['Schools', 'schools'];
         $crumbs[] = ['Delete', 'schools/delete'];
 
-        $this->view('schools.delete', [
-            'row' => $row,
-            'crumbs' => $crumbs
-        ]);
+        if(Auth::access('super_admin')) {
+            $this->view('schools.delete', [
+                'row' => $row,
+                'crumbs' => $crumbs
+            ]);
+        }else{
+            $this->view('access-denied');
+        }
     }
-
 }
